move example gql query into seperate directory to avoid clashes with prod ones
