+++ reconcile/gql_definitions_examples/fragments/vault_secret.gql
@@ -1,0 +1,8 @@
+# qenerate: plugin=pydantic_v1
+
+fragment VaultSecret on VaultSecret_v1 {
+    path
+    field
+    version
+    format
+}
