+++ reconcile/gql_definitions_examples/fragments/vault_secret.py
@@ -1,0 +1,27 @@
+"""
+Generated by qenerate plugin=pydantic_v1. DO NOT MODIFY MANUALLY!
+"""
+from typing import (  # noqa: F401 # pylint: disable=W0611
+    Any,
+    Callable,
+    Optional,
+    Union,
+)
+
+from pydantic import (  # noqa: F401 # pylint: disable=W0611
+    BaseModel,
+    Extra,
+    Field,
+    Json,
+)
+
+
+class VaultSecret(BaseModel):
+    path: str = Field(..., alias="path")
+    field: str = Field(..., alias="field")
+    version: Optional[int] = Field(..., alias="version")
+    q_format: Optional[str] = Field(..., alias="format")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
