+++ reconcile/gql_definitions_examples/fragments/vault_secret_version.py
