+++ reconcile/gql_definitions_examples/queries/jenkins.py
@@ -15,7 +15,7 @@ from pydantic import (  # noqa: F401 # pylint: disable=W0611
     Json,
 )
 
-from reconcile.gql_definitions.fragments.vault_secret import VaultSecret
+from reconcile.gql_definitions_examples.fragments.vault_secret import VaultSecret
 
 
 DEFINITION = """
