+++ reconcile/gql_definitions_examples/queries/ocm.py
@@ -15,9 +15,13 @@ from pydantic import (  # noqa: F401 # pylint: disable=W0611
     Json,
 )
 
-from reconcile.gql_definitions.fragments.vault_secret import VaultSecret
-from reconcile.gql_definitions.fragments.vault_secret_version import VaultSecretVersion
-from reconcile.gql_definitions.fragments.vault_secret_path import VaultSecretPath
+from reconcile.gql_definitions_examples.fragments.vault_secret import VaultSecret
+from reconcile.gql_definitions_examples.fragments.vault_secret_version import (
+    VaultSecretVersion,
+)
+from reconcile.gql_definitions_examples.fragments.vault_secret_path import (
+    VaultSecretPath,
+)
 
 
 DEFINITION = """
